+++ xray/backends/api.py
@@ -5,7 +5,7 @@ from io import BytesIO
 
 from .. import backends, conventions
 from .common import ArrayWriter
-from ..core.alignment import auto_combine
+from ..core.combine import auto_combine
 from ..core.utils import close_on_error, is_remote_uri
 from ..core.pycompat import basestring, OrderedDict, range