+++ xray/test/test_dask.py
@@ -145,7 +145,7 @@ class TestVariable(DaskTestCase):
         self.assertLazyAndIdentical(u, Variable.concat([v[:2], v[2:]], 'x'))
         self.assertLazyAndIdentical(u[:2], Variable.concat([v[0], v[1]], 'x'))
         self.assertLazyAndIdentical(
-            u[:3], Variable.concat([v[[0, 2]], v[[1]]], 'x', indexers=[[0, 2], [1]]))
+            u[:3], Variable.concat([v[[0, 2]], v[[1]]], 'x', positions=[[0, 2], [1]]))
 
     def test_missing_methods(self):
         v = self.lazy_var