+++ doc/whats-new.rst
@@ -27,6 +27,8 @@ Enhancements
 Bug fixes
 ~~~~~~~~~
 
+- Fix ``min``, ``max``, ``argmin`` and ``argmax``for arrays with string or
+  unicode types (:issue:`453`).
 - :py:func:`~xray.open_dataset` and :py:func:`~xray.open_mfdataset` support
   supplying chunks as a single integer.